Refactor out extensions from ConfigContainer
